Remove default case from property "value" marshalling
Previously the "value" field of a property object was automatically
filled with the raw uint64_t value by default. In particular, this
happened for blobs.

However I'd like to reserve the "value" field for blobs to contain
the base64-encoded blob contents. If we fill this with a uint64_t
value now, we won't be able to change it to a string later without
breaking consumers.

Additionally, this causes issues for future property types: they
would have their "value" field set to the uint64_t value too, making
it difficult to change it to a more appropriate type later.