Decode HDR_OUTPUT_METADATA
This was essential in debugging some HDR enablement code I'm working on.

In the pretty-print output, all fields are decoded to integers.
If someone deems it useful, I could put in the extra work to show enum
names and convert the fixed-point values to floats in the 0.0-1.0 range.

Im the json output, I deliberately flattened the primaries to separate
fields for the B, G and B components because I felt that the array
notation is less redable.